refactor partition builder and add basic.test
