+++ plugin/sycl/common/hist_util.cc
@@ -6,6 +6,7 @@
 #include <limits>
 #include <algorithm>
 
+#include "../data/gradient_index.h"
 #include "hist_util.h"
 
 #include <CL/sycl.hpp>
@@ -13,18 +14,6 @@
 namespace xgboost {
 namespace sycl {
 namespace common {
-
-uint32_t SearchBin(const bst_float* cut_values, const uint32_t* cut_ptrs, Entry const& e) {
-  auto beg = cut_ptrs[e.index];
-  auto end = cut_ptrs[e.index + 1];
-  const auto &values = cut_values;
-  auto it = std::upper_bound(cut_values + beg, cut_values + end, e.fvalue);
-  uint32_t idx = it - cut_values;
-  if (idx == end) {
-    idx -= 1;
-  }
-  return idx;
-}
 
 template <typename BinIdxType>
 void mergeSort(BinIdxType* begin, BinIdxType* end, BinIdxType* buf) {
@@ -46,138 +35,6 @@ void mergeSort(BinIdxType* begin, BinIdxType* end, BinIdxType* buf) {
       }
       for (size_t i = start; i < finish; i++) begin[i] = buf[i];
     }
-  }
-}
-
-template <typename BinIdxType>
-void GHistIndexMatrix::SetIndexData(::sycl::queue qu,
-                                    xgboost::common::Span<BinIdxType> index_data_span,
-                                    const DeviceMatrix &dmat_device,
-                                    size_t nbins,
-                                    size_t row_stride,
-                                    uint32_t* offsets) {
-  if (hit_count.size() == 0) return;
-  const xgboost::Entry *data_ptr = dmat_device.data.DataConst();
-  const bst_row_t *offset_vec = dmat_device.row_ptr.DataConst();
-  const size_t num_rows = dmat_device.row_ptr.Size() - 1;
-  BinIdxType* index_data = index_data_span.data();
-  const bst_float* cut_values = cut_device.Values().DataConst();
-  const uint32_t* cut_ptrs = cut_device.Ptrs().DataConst();
-  ::sycl::buffer<size_t, 1> hit_count_buf(hit_count.data(), hit_count.size());
-
-  USMVector<BinIdxType> sort_buf(&qu, num_rows * row_stride);
-  BinIdxType* sort_data = sort_buf.Data();
-
-  qu.submit([&](::sycl::handler& cgh) {
-    auto hit_count_acc = hit_count_buf.template get_access<::sycl::access::mode::atomic>(cgh);
-    cgh.parallel_for<>(::sycl::range<1>(num_rows), [=](::sycl::item<1> pid) {
-      const size_t i = pid.get_id(0);
-      const size_t ibegin = offset_vec[i];
-      const size_t iend = offset_vec[i + 1];
-      const size_t size = iend - ibegin;
-      const size_t start = i * row_stride;
-      for (bst_uint j = 0; j < size; ++j) {
-        uint32_t idx = SearchBin(cut_values, cut_ptrs, data_ptr[ibegin + j]);
-        index_data[start + j] = offsets ? idx - offsets[j] : idx;
-        ::sycl::atomic_fetch_add<size_t>(hit_count_acc[idx], 1);
-      }
-      if (!offsets) {
-        // Sparse case only
-        mergeSort<BinIdxType>(index_data + start, index_data + start + size, sort_data + start);
-        for (bst_uint j = size; j < row_stride; ++j) {
-          index_data[start + j] = nbins;
-        }
-      }
-    });
-  }).wait();
-}
-
-void GHistIndexMatrix::ResizeIndex(const size_t n_offsets,
-                                         const size_t n_index,
-                                         const bool isDense) {
-  if ((max_num_bins - 1 <= static_cast<int>(std::numeric_limits<uint8_t>::max())) && isDense) {
-    index.SetBinTypeSize(BinTypeSize::kUint8BinsTypeSize);
-    index.Resize((sizeof(uint8_t)) * n_index);
-  } else if ((max_num_bins - 1 > static_cast<int>(std::numeric_limits<uint8_t>::max())  &&
-    max_num_bins - 1 <= static_cast<int>(std::numeric_limits<uint16_t>::max())) && isDense) {
-    index.SetBinTypeSize(BinTypeSize::kUint16BinsTypeSize);
-    index.Resize((sizeof(uint16_t)) * n_index);
-  } else {
-    index.SetBinTypeSize(BinTypeSize::kUint32BinsTypeSize);
-    index.Resize((sizeof(uint32_t)) * n_index);
-  }
-}
-
-void GHistIndexMatrix::Init(::sycl::queue qu,
-                            Context const * ctx,
-                            const DeviceMatrix& p_fmat_device,
-                            int max_bins) {
-  nfeatures = p_fmat_device.p_mat->Info().num_col_;
-
-  cut = xgboost::common::SketchOnDMatrix(ctx, p_fmat_device.p_mat, max_bins);
-  cut_device.Init(qu, cut);
-
-  max_num_bins = max_bins;
-  const uint32_t nbins = cut.Ptrs().back();
-  this->nbins = nbins;
-  hit_count.resize(nbins, 0);
-
-  this->p_fmat = p_fmat_device.p_mat;
-  const bool isDense = p_fmat_device.p_mat->IsDense();
-  this->isDense_ = isDense;
-
-  row_ptr = std::vector<size_t>(p_fmat_device.row_ptr.Begin(), p_fmat_device.row_ptr.End());
-  row_ptr_device = p_fmat_device.row_ptr;
-
-  index.setQueue(qu);
-
-  row_stride = 0;
-  for (const auto& batch : p_fmat_device.p_mat->GetBatches<SparsePage>()) {
-    const auto& row_offset = batch.offset.ConstHostVector();
-    for (auto i = 1ull; i < row_offset.size(); i++) {
-      row_stride = std::max(row_stride, static_cast<size_t>(row_offset[i] - row_offset[i - 1]));
-    }
-  }
-
-  const size_t n_offsets = cut.Ptrs().size() - 1;
-  const size_t n_rows = p_fmat_device.row_ptr.Size() - 1;
-  const size_t n_index = n_rows * row_stride;
-  ResizeIndex(n_offsets, n_index, isDense);
-
-  CHECK_GT(cut.Values().size(), 0U);
-
-  uint32_t* offsets = nullptr;
-  if (isDense) {
-    index.ResizeOffset(n_offsets);
-    offsets = index.Offset();
-    qu.memcpy(offsets, cut.Ptrs().data(), sizeof(uint32_t) * n_offsets).wait_and_throw();
-    // for (size_t i = 0; i < n_offsets; ++i) {
-    //   offsets[i] = cut.Ptrs()[i];
-    // }
-  }
-
-  if (isDense) {
-    BinTypeSize curent_bin_size = index.GetBinTypeSize();
-    if (curent_bin_size == BinTypeSize::kUint8BinsTypeSize) {
-      xgboost::common::Span<uint8_t> index_data_span = {index.data<uint8_t>(),
-                                                        n_index};
-      SetIndexData(qu, index_data_span, p_fmat_device, nbins, row_stride, offsets);
-
-    } else if (curent_bin_size == BinTypeSize::kUint16BinsTypeSize) {
-      xgboost::common::Span<uint16_t> index_data_span = {index.data<uint16_t>(),
-                                                         n_index};
-      SetIndexData(qu, index_data_span, p_fmat_device, nbins, row_stride, offsets);
-    } else {
-      CHECK_EQ(curent_bin_size, BinTypeSize::kUint32BinsTypeSize);
-      xgboost::common::Span<uint32_t> index_data_span = {index.data<uint32_t>(),
-                                                         n_index};
-      SetIndexData(qu, index_data_span, p_fmat_device, nbins, row_stride, offsets);
-    }
-  /* For sparse DMatrix we have to store index of feature for each bin
-     in index field to chose right offset. So offset is nullptr and index is not reduced */
-  } else {
-    xgboost::common::Span<uint32_t> index_data_span = {index.data<uint32_t>(), n_index};
-    SetIndexData(qu, index_data_span, p_fmat_device, nbins, row_stride, offsets);
   }
 }
